# docs/swagger.yaml
basePath: /MatchingApp
host: localhost:8080
info:
  contact: {}
  description: finding your spotify match
  title: Matching App
paths:
...
  /createUser:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided username, password, and email.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
...
  /loginUser:
    post:
      consumes:
      - application/json
      description: Authenticate a user with the provided username and password.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            type: string
        "400":
          description: Invalid username or password
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login a user
      tags:
      - users
...
swagger: "2.0"